pyproject.toml
[tool.poetry]
name = "solana-dataset-kaggle-auto-updater"
version = "0.1.0"
description = "Automated Solana historical data fetcher and Kaggle dataset updater"
authors = ["Novandra Anugrah <your.email@example.com>"]
readme = "README.md"
packages = [{include = "solana_dataset_kaggle_auto_updater"}]

[tool.poetry.dependencies]
python = "^3.8.1"
pandas = "^1.5.0"
python-binance = "^1.0.19"
python-dotenv = "^1.0.0"
kaggle = "^1.5.16"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
update-solana = "solana_dataset_kaggle_auto_updater.kaggle_update_solana:main"
fetch-full-solana = "fetch_full_sol_data:main" 